// fe-spa-durban-admin-develop-main/src/modules/CashbackRules/models/CashBack.model.ts
export type CashBack = {
  cashBackRulesName: string;
  howMuchCashback: string;
  cashBackDate: string;
  status: string;
  _id: string;
  isActive: boolean;
};

export type CashBackFormValues = {
  // outlets: any;
  cashBackRulesName: string;
  howMuchCashback: any;
  cashBackDate: any;
  cashBackEndDate: any;
  serviceId: any;
};